// src/utils/types.ts
export interface Invoice {
  id: string;
  createdAt: string;
  paymentDue: string;
  description: string;
  paymentTerms: number;
  clientName: string;
  clientEmail: string;
  status: InvoiceStatus;
  senderAddress: {
    street: string;
    city: string;
    postCode: string;
    country: string;
  };
  clientAddress: {
    street: string;
    city: string;
    postCode: string;
    country: string;
  };
  items: InvoiceItem[];
  total: number;
}

export type InvoiceStatus = "paid" | "pending" | "draft";

export type InvoiceItem = {
  name: string;
  quantity: number;
  price: number;
  total: number;
};